Survey Form
Survey form where all the fields are required. If all the fields are not set, then the error message throws. If all values are filled and pressed Enter or Submit, pop up will come with the Label and values. On closing this pop up all the values will be reseted.